[Debug] Add mod code

+++ mod/plr.h
@@ -20,7 +20,7 @@ struct line {
 class Segment {
 public:
     Segment(uint64_t _x, double _k, double _b, uint64_t _x2) : x(_x), k(_k), b(_b), x2(_x2) {}
-    uint64_t x;
+    double x;
     double k;
     double b;
     uint64_t x2;
